+++ src/hittable.h
@@ -11,7 +11,7 @@ typedef struct hit_record_t
 	point3_t	p;
 	vec3_t		normal;
 	material_t* mat;
-	f64			t;
+	f32			t;
 	b32			front_face;
 } hit_record_t;
